backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
qdrant-client==1.7.0
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
markdown==3.5.1
beautifulsoup4==4.12.2
aiofiles==23.2.1
httpx==0.25.2
google-generativeai==0.3.2